This adds the option to get a properly configured gRPC connection to the emulator in which this test is currently running.
The gRPC endpoint will be configured to:

- Use TLS if needed.
- Set a proper authorization header with token if needed.

The configuration is obtained from the arguments provided by the test runner.
The following arguments are used:

  - grpc.port: The port where the emulator is hosting the gRPC services
  - grpc.token: JWT used to authenticate
  - grpc.key (optional): Path to private key used to configure tls
  - grpc.cer (optional): Path to the public cert used to configure tls
  - grpc.ca (optional): Certificate authority used to configure tls

The optional parameters require the injection of the actual pem and cert files to
the android file system. They should only be visible to the current process under
test

PiperOrigin-RevId: 411868032